pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DiscordianAI"
version = "0.2.9"
description = "A Discord bot that uses OpenAI's GPT API to generate responses."
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "johndotpub", email = "github@john.pub"}]
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.6.4",
    "openai>=2.24.0",
    "websockets>=16.0",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "httpx[http2]>=0.28.1",
    "h2>=4.3.0"
]
urls = {"Homepage" = "https://github.com/johndotpub/DiscordianAI"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[project.scripts]
discordianai = "src.main:main"

[project.optional-dependencies]
dev = [
    "black>=26.1.0", 
    "ruff>=0.15.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0", 
    "pytest-xdist[psutil]>=3.8.0",  # For parallel test execution
    "coverage[toml]>=7.13.4",
    "tox>=4.46.3",
    "pip-audit>=2.10.0",  # Dependency vulnerability scanning
    "pre-commit>=4.5.1",  # Pre-commit hooks framework
    "detect-secrets>=1.5.0",  # Secret detection for pre-commit
]

[tool.pytest.ini_options]
addopts = "-ra -q"
filterwarnings = [
    "ignore:'audioop' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
]
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 99
extend-exclude = '''
/(
    \.git
  | __pycache__
  | build
  | dist
  | \.tox
  | \.venv
  | venv
  | \.egg-info
  | \.pytest_cache
  | \.coverage.*
  | htmlcov
  | .*\.log
  | scripts
)/
'''

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/__init__.py",
    "*/tests/*",
    "*/test_*"
]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.ruff]
target-version = "py310"
line-length = 99
extend-exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist", 
    ".tox",
    ".venv",
    "venv",
    "*.egg-info",
    ".pytest_cache",
    ".coverage*",
    "htmlcov",
    "*.log",
    "scripts",
]

[tool.ruff.lint]
# Enable comprehensive rule sets - modern replacement for flake8
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "BLE",    # flake8-blind-except
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "TD",     # flake8-todos
    "FIX",    # flake8-fixme
    "ERA",    # eradicate
    "PD",     # pandas-vet
    "PGH",    # pygrep-hooks
    "PL",     # Pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "NPY",    # NumPy-specific rules
    "PERF",   # Perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "RUF",    # Ruff-specific rules
]



# Rules to ignore (equivalent to flake8 ignore)
ignore = [
    "S101",   # Use of assert detected (OK in tests)
]

[tool.ruff.lint.mccabe]
# Complexity limit (equivalent to flake8 --max-complexity)
max-complexity = 16

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings
convention = "google"

[tool.ruff.lint.isort]
# Configure isort behavior within ruff
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.per-file-ignores]
# Test files can have additional flexibility
"tests/*.py" = [
    "D",       # Skip docstring requirements in tests
    "S101",    # Allow assert in tests
    "PLR2004", # Allow magic values in tests
    "ARG001",  # Allow unused function arguments in tests
    "FBT001",  # Allow boolean trap in tests
    "PLC0415", # Allow local imports in tests
    "ARG004",  # Allow unused *args/**kwargs in mocks
    "PT011",   # Broad raises in tests
    "PT012",   # Complex raises block in tests
    "SLF001",  # Private access in tests
    "SIM117",  # Nested context managers in tests
    "PTH",     # Pathlib preference in tests
    "B007",    # Unused loop var in tests
    "ASYNC251",# time.sleep in async tests
    "TRY002",  # Use of Exception in tests
]

# Loosen specific modules to reduce noisy warnings without changing behavior
# Keep core modules strict; no unnecessary ignores remain

